Extract SQL queries
