+++ cypher-to-sql/src/main/resources/sql-queries/purge.sql
@@ -1,0 +1,3 @@
+DROP SCHEMA IF EXISTS PUBLIC CASCADE;
+CREATE SCHEMA PUBLIC AUTHORIZATION "postgres";
+GRANT ALL ON SCHEMA public TO public;
